extracts avatar into it's own helper
Also moves tests to specs
